moved configuration to services
