+++ src/app/shared/services/configuration.data.ts
