+++ src/app/shared/services/configuration.service.ts
